Use a common enum for piece values
Move BLACK, WHITE, and EMPTY from the AbaloneBoard class to the package abalone, since they are a pretty common thing, shared by both the view and the model. Other people might think it was perfectly fine to keep it in AbaloneBoard.

The BIG thing this fixed was removing a ton of mysterious unnamed constants such as the default value for val in the Hex constructor was 0, not EMPTY. You were just getting lucky before that EMPTY happened to be declared as 0. I ran into all sorts of bugs when I changed the values of BLACK, WHITE, etc. To demonstrate and test this I set the values of the enums to random numbers, and it still better work.